Move IR-related virtual files to `linker.ioir.*`.
In addition, make sure that all files in the `tools` belong to one
of the three subpackages `io`, `logging` and `linker`.
