+++ sbt-plugin/src/main/scala/org/scalajs/sbtplugin/ScalaJSPluginInternal.scala
@@ -13,7 +13,9 @@ import sbt.complete.DefaultParsers._
 import org.portablescala.sbtplatformdeps.PlatformDepsPlugin.autoImport._
 
 import org.scalajs.core.tools.io.{IO => _, _}
+
 import org.scalajs.core.tools.linker._
+import org.scalajs.core.tools.linker.irio._
 import org.scalajs.core.tools.linker.standard._
 
 import org.scalajs.jsenv._
