+++ tools/js/src/test/scala/org/scalajs/core/tools/linker/testutils/Platform.scala
@@ -1,6 +1,6 @@
 package org.scalajs.core.tools.linker.testutils
 
-import org.scalajs.core.tools.io._
+import org.scalajs.core.tools.linker.irio._
 
 object Platform {
   def loadJar(path: String): ScalaJSIRContainer =
