+++ tools/jvm/src/test/scala/org/scalajs/core/tools/linker/testutils/Platform.scala
@@ -2,7 +2,7 @@ package org.scalajs.core.tools.linker.testutils
 
 import java.io.File
 
-import org.scalajs.core.tools.io._
+import org.scalajs.core.tools.linker.irio._
 
 object Platform {
   def loadJar(path: String): ScalaJSIRContainer =
