+++ tools/shared/src/main/scala/org/scalajs/core/tools/linker/irio/MemIRFiles.scala
@@ -6,7 +6,9 @@
 **                          |/____/                                     **
 \*                                                                      */
 
-package org.scalajs.core.tools.io
+package org.scalajs.core.tools.linker.irio
+
+import org.scalajs.core.tools.io._
 
 /** A simple in-memory mutable virtual serialized Scala.js IR file. */
 class MemVirtualSerializedScalaJSIRFile(p: String)
